user Sequence completed

+++ MensFashion/src/main/java/com/mensfashion/entities/UserEntity.java
@@ -25,6 +25,6 @@ public class UserEntity {
 	private String name;
 	private String email;
 	private String password;
-	private BigDecimal phoneNum;
+	private long phoneNum;
 	
 }
